requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
flashtext==2.7
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
magic-filter==1.0.12
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
narwhals==2.1.2
numpy==2.3.2
opensearch-py==2.5.0
orjson==3.11.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.2
pillow==10.4.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pyarrow==21.0.0
pyasn1==0.6.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==22.3
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rich-toolkit==0.15.0
rpds-py==0.27.0
rsa==4.9.1
selenium==4.35.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
streamlit==1.37.1
telegram==0.0.1
Telethon==1.40.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
